Fix no "img" and no "collectible" tag bug
/card had a bug that would raise an exception if all of the returned
cards had no "img" and no "collectible" tag. Since output was
initially set to "", if this were the case the function would return ""
and the bot would attempt to post an empty message. Now, if there is a
non-collectible card that has an image, it will be posted instead; next,
if none of the returned cards have an "img" tag, output's default text
(an explanation of the situation) will be returned.