Implement Wednesday posts
weekday_timers.py is a new file with a function that calculates the time
until the next specified weekday, which is represented in the weekday.py
enum file. It currently only supports until midnight on that day, but
could (hypothetically) easily be expanded to support variable hours,
minutes, etc. main.py uses that function to sleep a thread until the
next Wednesday at midnight, when it then posts a meme in the specified
chat channel.

Note: Much of the work that went into this commit comes from my commits
in the TCSS360 repository. Some of the iterations at making this work
come from there.